requirements.txt
gunicorn
Flask
setuptools
mongoengine
dnspython
flask-cors
requests
marshmallow
pika
pika-stubs
python-dotenv

-e .
